+++ HISTORY.md
@@ -5,6 +5,7 @@
 
 - Fixed the `babble` module always sending messages via it's own internal
   messagebus instead of the agents send/receive function.
+- Improved error handling for sending messages.
 
 
 ## 2014-10-23, version 0.4.2